% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benford.R
\name{benford}
\alias{benford}
\title{This function calculates Benford's Low expected versus observed values.}
\usage{
benford(x, main = "Variable of Interest", legend = c("Benford's Law",
  "Variable of Interest"))
}
\arguments{
\item{x}{is a vector with numeric values}

\item{main}{is a main title with default value of "Variable of Interest"}

\item{legend}{is a legends in the plot.}
}
\value{
plot of expectied and observed values of 1 to 9 intigers.
}
\description{
This function calculates Benford's Low expected versus observed values.
}
