% Generated by roxygen2 (4.0.2): do not edit by hand
\name{createDummies}
\alias{createDummies}
\title{Create dummy variables for each level of a categorical variable}
\usage{
createDummies(x, df, keepNAs = TRUE)
}
\arguments{
\item{x}{is a character vector with the name or names of the columnes in dataframe to convert to dumy values}

\item{df}{is a data.frame with new dummy variables.}

\item{keepNAs}{is logical vector for keeping or removing NA values.}
}
\value{
returns data.frame wiht dummy variables.
}
\description{
Create dummy variables for each level of a categorical variable
}

