% Generated by roxygen2 (4.0.2): do not edit by hand
\name{downloadFiles}
\alias{downloadFiles}
\title{Donwnload files function for downloading and storing data file in .csv format
adapted from Machine Learning Class}
\usage{
downloadFiles(dataURL = "", destF = "t.csv")
}
\arguments{
\item{dataURL}{url of the data to be downloaded}

\item{destF}{destination file to be saved in csv format.}
}
\value{
returns localy stored file, if files exists it will show message "Data alreade donwloaded".
}
\description{
Donwnload files function for downloading and storing data file in .csv format
adapted from Machine Learning Class
}

