% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggkmTable.R
\name{ggkmTable}
\alias{ggkmTable}
\title{Create a Kaplan-Meier plot using ggplot2}
\usage{
ggkmTable(sfit, table = TRUE, returns = FALSE, xlabs = "Time",
  ylabs = "survival probability", ystratalabs = NULL, ystrataname = NULL,
  timeby = 100, main = "", pval = TRUE, ...)
}
\arguments{
\item{sfit}{a \code{\link[survival]{survfit}} object}

\item{table}{logical: Create a table graphic below the K-M plot, indicating at-risk numbers?}

\item{returns}{logical: if \code{TRUE}, return an arrangeGrob object}

\item{xlabs}{x-axis label}

\item{ylabs}{y-axis label}

\item{ystratalabs}{The strata labels. \code{Default = levels(summary(sfit)$strata)}}

\item{ystrataname}{The legend name. Default = “Strata”}

\item{timeby}{numeric: control the granularity along the time-axis}

\item{main}{plot title}

\item{pval}{logical: add the pvalue to the plot?}
}
\value{
a ggplot is made. if return=TRUE, then an arrangeGlob object
is returned
}
\description{
Create a Kaplan-Meier plot using ggplot2
}
\examples{
\dontrun{
data(colon)
 fit <- survfit(Surv(time,status)~rx, data=colon)
 ggkm(fit, timeby=500)
}
}
\author{
Abhijit Dasgupta with contributions by Gil Tomas
\url{http://statbandit.wordpress.com/2011/03/08/an-enhanced-kaplan-meier-plot/}
}
