% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guessDateFormat.R
\name{guessDateFormat}
\alias{guessDateFormat}
\title{FUNCTION guessDateFormat @x vector of character dates/datetimes @returnDates return
actual dates rather than format convert character datetime to POSIXlt datetime, or at # least guess the format such that you could convert to datetime}
\usage{
guessDateFormat(x, returnDates = FALSE, tzone = "")
}
\arguments{
\item{x}{date object}

\item{tzone}{time zone default is not nececcary}
}
\description{
FUNCTION guessDateFormat @x vector of character dates/datetimes @returnDates return
actual dates rather than format convert character datetime to POSIXlt datetime, or at # least guess the format such that you could convert to datetime
}
