% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lm_diag}
\alias{lm_diag}
\title{linear model function diagnostic with custom output}
\usage{
lm_diag(formula, dataset)
}
\arguments{
\item{formula}{for linear model with y outcome ~ relation and covariates divided with + and interacinos with : and/or *}

\item{dataset}{is the data for lm model}
}
\value{
returns three diagnostics graphs plotted in base plotting systime.
}
\description{
linear model function diagnostic with custom output
}

