% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_ci.R
\name{m.ci}
\alias{m.ci}
\title{confidence intervals for latex documents.}
\usage{
m.ci(x, na.rm = TRUE)
}
\arguments{
\item{x}{numeric vector for mean and range calculation}

\item{na.rm}{logical value for removing or not the NA values.}
}
\value{
returns mean and standard deviation for latex envirment.
}
\description{
confidence intervals for latex documents.
}
