% Generated by roxygen2 (4.0.2): do not edit by hand
\name{m.rg}
\alias{m.rg}
\title{median and range for latex documents.}
\usage{
m.rg(x, na.rm = TRUE)
}
\arguments{
\item{x}{numeric vector for mean and range calculation}
}
\value{
returns mean and range (0.25 percentile - 0.75 percentile) of the vector for latex envirment.
}
\description{
median and range for latex documents.
}

