% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_sd.R
\name{m.sd}
\alias{m.sd}
\title{mean and standard deviation for latex documents.}
\usage{
m.sd(x, na.rm = TRUE)
}
\arguments{
\item{x}{numeric vector for mean and range calculation}

\item{na.rm}{logical value for removing or not the NA values.}
}
\value{
returns mean and standard deviation for latex envirment.
}
\description{
mean and standard deviation for latex documents.
}
