% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_df.R
\name{missing_df}
\alias{missing_df}
\title{A function that plots missingness
requires `reshape2`}
\usage{
missing_df(x)
}
\arguments{
\item{x}{dataframe class object}
}
\description{
A function that plots missingness
requires `reshape2`
}
