% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msmr.R
\name{msmr}
\alias{msmr}
\title{Estimating the mean and standard deviation from the median and the range
From statmd blog}
\usage{
msmr(a, m, b, n)
}
\arguments{
\item{a}{min of the data}

\item{m}{median of the data}

\item{b}{max of the data}

\item{n}{sample of size}
}
\description{
Estimating the mean and standard deviation from the median and the range
From statmd blog
}
