% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiLib.R
\name{multiLibrary}
\alias{multiLibrary}
\title{transform code below into the multiLibrary function}
\usage{
multiLibrary(libraries)
}
\arguments{
\item{libraries}{character vector with library name or names to load.}
}
\value{
lodes libraries into curnt session of R
}
\description{
transform code below into the multiLibrary function
}
