% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.R
\name{p}
\alias{p}
\title{Calculate percentage in a numeric vector}
\usage{
p(x, level = 1, f = FALSE, na.rm = TRUE, ...)
}
\arguments{
\item{x}{is a numeric vector}

\item{level}{is a level of the fator ocharacter variable}

\item{na.rm}{is remouval of NA values and set defaoul to TRUE}
}
\value{
numeric value expressed in percents
}
\description{
Calculate percentage in a numeric vector
}
