% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perc_n.R
\name{perc_n}
\alias{perc_n}
\title{Calculate percentage in a numeric vector}
\usage{
perc_n(x, na.rm = TRUE, ref = 1, ...)
}
\arguments{
\item{x}{is a numeric vector}

\item{na.rm}{is remouval of NA values and set defaoul to TRUE}
}
\value{
numeric value expressed in percents
}
\description{
Calculate percentage in a numeric vector
}
