% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.pval.latex}
\alias{print.pval.latex}
\title{thresholded p-value 0.01 for latex implementation}
\usage{
\method{print}{pval.latex}(pval)
}
\arguments{
\item{p-value}{object}
}
\value{
returns pvalue in latex environment.
}
\description{
thresholded p-value 0.01 for latex implementation
}

