% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pval_latex}
\alias{pval_latex}
\title{Thresholded p-value 0.01 for latex implementation}
\usage{
pval_latex(pval)
}
\arguments{
\item{pval}{object}
}
\value{
returns pvalue in latex environment.
}
\description{
Thresholded p-value 0.01 for latex implementation
}

