% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pval_plot}
\alias{pval_plot}
\title{thresholded p-value for ploting purposes}
\usage{
pval_plot(pval)
}
\arguments{
\item{pval}{object}
}
\value{
returns pvalue in base ploting system environment.
}
\description{
thresholded p-value for ploting purposes
}

