\name{raffdoc-package}
\alias{raffdoc-package}
\alias{raffdoc}
\docType{package}
\title{
Package for short and easy fuctions customized by raffdoc
}
\description{
This package can be used in latex and rmd enviroments for different useful fuctions.
}
\details{
\tabular{ll}{
Package: \tab raffdoc\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-09-19\cr
License: \tab What license is it under?\cr
}
It is easy to use, documentation si provided.
}
\author{
raffdoc

Maintainer: Who to complain to <margaryan@ftgm.it>

}
\references{
Margaryan et al. 2014.
}

\keyword{ package }
\seealso{

}
\examples{

}
