% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rows_all_missing.R
\name{rows_all_missing}
\alias{rows_all_missing}
\title{Calculate percentage in a numeric vector}
\usage{
rows_all_missing(df)
}
\arguments{
\item{df}{is a dataframe object}
}
\value{
numeric value expressed in percents
}
\description{
Calculate percentage in a numeric vector
}
