% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_info.R
\name{type_info}
\alias{type_info}
\title{median and range for latex documents.}
\usage{
type_info(x)
}
\arguments{
\item{x}{object in wich we want to know the properties}
}
\value{
returns list of propertes in the object
}
\description{
median and range for latex documents.
}
