% Generated by roxygen2 (4.0.2): do not edit by hand
\name{hist_wnc}
\alias{hist_wnc}
\title{histogram wiht base plotting system with normal curve
addopted from http://www.statmethods.net/graphs/density.html}
\usage{
hist_wnc(x, breaks = 24)
}
\arguments{
\item{x}{numeric vector wiht values to plot}

\item{breaks}{count of breaks in the histogram.}
}
\value{
histogram wiht base plotting sistem
}
\description{
histogram wiht base plotting system with normal curve
addopted from http://www.statmethods.net/graphs/density.html
}

