% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_rg.R
\name{m.rg}
\alias{m.rg}
\title{median and range for latex documents.}
\usage{
m.rg(x, na.rm = TRUE)
}
\arguments{
\item{x}{numeric vector for median and range calculation}

\item{na.rm}{logical value for removing or not the NA values.}
}
\value{
returns median and range (0.25 percentile - 0.75 percentile) of the vector for latex envirment.
}
\description{
median and range for latex documents.
}
