% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_plot.R
\name{pval_plot}
\alias{pval_plot}
\title{thresholded p-value for ploting purposes}
\usage{
pval_plot(pval)
}
\arguments{
\item{pval}{object}
}
\value{
returns pvalue in base ploting system environment.
}
\description{
thresholded p-value for ploting purposes
}
